Add BUILD_ALL_EXTENSIONS ref website instructions
If CMake is called in a platform that supports Vulkan, but the necessary
Vulkan requirements are not in place, building the documentation fails.
It should be possible to build the documentation without having the
Vulkan requirements in place. That is possible by setting
BUILD_ALL_EXTENSIONS to FALSE. Add that to the tutorial website build
instructions.